server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
# numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
cloudinary==1.36.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
